code array
yes array

+++ src/_01_Simple_Array_Algorithms/_00_1D_Array_Methods.java
@@ -15,10 +15,11 @@ for (int i = 0; i < values.length; i++) {
     //2. Complete the method so that it returns the average of all
     //   of the integers in the array being passed in
     public static double averageIntArray(int[] values) {
-    	int avg = 0;
+    	double avg = 0;
     	for (int i = 0; i < values.length; i++) {
-    		avg = avg +values[i];
+    		avg = avg + values[i];
     	}
+    	
     	avg = avg/values.length;
     	
         return avg;
